Updated Delete / Create Comment
I added incrementing when creating a comment as well as decrementing when the comment is deleted. I didnt want to change much of the persons code so I really just changed what was in db folder. Also there were two routes for deleting comments and I got rid of one.