Feed Endpoint - Update
Made a change to the way we fetch profile pic because it is written as profile_picture in the user model and is in the users subdocument so I added the .user_info. I notice that when I test on postman. I dont see the profile picture in the object but I believe that is because our db currently does not have any pictures. As the pictures are not required on the diagram we have.